tests: socat: refactor bittorrent in socat generic test
Now we can reuse this for all sorts of payloads
